Fix CI: HeartRateRecord.BPM_AVG aggrega a Long, non Double
Unico errore di compilazione sul nuovo HealthConnectManager: il resto
della superficie API di Health Connect ipotizzata (client, permessi,
AggregateRequest, TimeRangeFilter, SleepSessionRecord.SLEEP_DURATION_TOTAL)
ha compilato al primo colpo. WeeklyHealthAverages.avgHeartRateBpm
Double? -> Long?, formattazione in DashboardScreen.kt corretta di
conseguenza (roundToInt() non esiste su Long).

+++ app/src/main/java/com/simone/jarvismobile/health/HealthConnectManager.kt
@@ -72,7 +72,10 @@ class HealthConnectManager @Inject constructor(
     }
 
     data class WeeklyHealthAverages(
-        val avgHeartRateBpm: Double?,
+        // BPM_AVG aggregates to a Long (§ fix CI: "Argument type mismatch:
+        // actual type is 'Long?', but 'Double?' was expected" — the library's
+        // real signature, not the Double this was first guessed as).
+        val avgHeartRateBpm: Long?,
         val avgSleepPerNight: Duration?,
     )
